chore: refactor config to its own package
