+++ pkg/config/hooks.go
@@ -1,0 +1,6 @@
+package config
+
+// Hooks a list of commands to execute after the templating is done.
+type Hooks struct {
+	Post []string `yaml:"post"`
+}
