+++ pkg/run_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/ilaif/goplicate/pkg"
 	"github.com/ilaif/goplicate/pkg/cmd/testutils"
+	"github.com/ilaif/goplicate/pkg/config"
 	"github.com/ilaif/goplicate/pkg/mocks"
 )
 
@@ -16,11 +17,11 @@ func TestRun_Error_SyncingToNonExistentFile(t *testing.T) {
 
 	defer testutils.PrepareWorkdir(t, "../examples/sync-initial", ".")()
 
-	config := &pkg.ProjectConfig{
-		Targets: []pkg.Target{
+	cfg := &config.ProjectConfig{
+		Targets: []config.Target{
 			{
 				Path:        "config.yaml",
-				Source:      pkg.Source{Path: "./shared/config.yaml"},
+				Source:      config.Source{Path: "./shared/config.yaml"},
 				SyncInitial: false,
 			},
 		},
@@ -28,7 +29,7 @@ func TestRun_Error_SyncingToNonExistentFile(t *testing.T) {
 	cloner := &mocks.ClonerMock{}
 	opts := pkg.NewRunOpts(false, true, false, false, false, false, "")
 
-	r.ErrorContains(pkg.Run(context.TODO(), config, cloner, opts), "Failed to read file")
+	r.ErrorContains(pkg.Run(context.TODO(), cfg, cloner, opts), "Failed to read file")
 }
 
 func TestRun_Success_SyncingToNonExistentFile_WithSyncInitial(t *testing.T) {
@@ -36,11 +37,11 @@ func TestRun_Success_SyncingToNonExistentFile_WithSyncInitial(t *testing.T) {
 
 	defer testutils.PrepareWorkdir(t, "../examples/sync-initial", ".")()
 
-	config := &pkg.ProjectConfig{
-		Targets: []pkg.Target{
+	cfg := &config.ProjectConfig{
+		Targets: []config.Target{
 			{
 				Path:        "config.yaml",
-				Source:      pkg.Source{Path: "./shared/config.yaml"},
+				Source:      config.Source{Path: "./shared/config.yaml"},
 				SyncInitial: true,
 			},
 		},
@@ -48,7 +49,7 @@ func TestRun_Success_SyncingToNonExistentFile_WithSyncInitial(t *testing.T) {
 	cloner := &mocks.ClonerMock{}
 	opts := pkg.NewRunOpts(false, true, false, false, false, false, "")
 
-	r.NoError(pkg.Run(context.TODO(), config, cloner, opts))
+	r.NoError(pkg.Run(context.TODO(), cfg, cloner, opts))
 
 	testutils.RequireFileContains(r, "config.yaml", "key: value")
 }
