+++ pkg/source.go
@@ -7,12 +7,13 @@ import (
 	"github.com/caarlos0/log"
 	"github.com/pkg/errors"
 
+	"github.com/ilaif/goplicate/pkg/config"
 	"github.com/ilaif/goplicate/pkg/git"
 )
 
 // ResolveSourcePath given a source, resolves it by cloning the repository (if applicable)
 // and returning the directory of the source.
-func ResolveSourcePath(ctx context.Context, source Source, workdir string, cloner git.Cloner) (string, error) {
+func ResolveSourcePath(ctx context.Context, source config.Source, workdir string, cloner git.Cloner) (string, error) {
 	log.Debugf("Resolving path of source '%s'", source.String())
 
 	var err error
